+++ test/test_cmath.cpp
@@ -1474,316 +1474,6 @@ void test_nan()
 }
 #endif
 
-template <typename T>
-void test_erf()
-{
-    std::uniform_int_distribution<int> dist(1, 1);
-
-    BOOST_TEST(isnan(erf(std::numeric_limits<T>::quiet_NaN()) * dist(rng)));
-    BOOST_TEST_EQ(erf(T(0 * dist(rng))), T{0});
-    BOOST_TEST_EQ(erf(std::numeric_limits<T>::infinity() * dist(rng)), T{1});
-    BOOST_TEST_EQ(erf(-std::numeric_limits<T>::infinity() * dist(rng)), T{-1});
-
-    // Small values
-    using float_type = std::conditional_t<std::is_same<T, decimal32>::value, float, double>;
-
-    std::uniform_real_distribution<float_type> smallest_vals(1e-15, 1e-10);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {smallest_vals(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> second_smallest(1e-10, 0.5);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {second_smallest(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> erfc_smallest(0.5, 1.5);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {erfc_smallest(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> erfc_middle(1.5, 2.5);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {erfc_middle(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> erfc_large(2.5, 4.5);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {erfc_large(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> catch_all(4.5, 110.0);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {catch_all(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    // Negative values
-    std::uniform_real_distribution<float_type> negative(-2, 0.0);
-    for (std::size_t i {}; i < N / 8; ++i)
-    {
-        const auto val {negative(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "  Val: " << val
-                      << "\nFloat: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    // Underflow case
-    BOOST_TEST_EQ(erf(T{120}), T{1} * dist(rng));
-}
-
-template <typename T>
-void test_erfc()
-{
-    std::uniform_int_distribution<int> dist(1, 1);
-
-    BOOST_TEST(isnan(erfc(std::numeric_limits<T>::quiet_NaN() * dist(rng))));
-    BOOST_TEST_EQ(erfc(std::numeric_limits<T>::infinity() * dist(rng)), T{0});
-    BOOST_TEST_EQ(erfc(-std::numeric_limits<T>::infinity() * dist(rng)), T{2});
-
-    // Small values
-    using float_type = std::conditional_t<std::is_same<T, decimal32>::value, float, double>;
-
-    std::uniform_real_distribution<float_type> smallest_vals(1e-15, 1e-10);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {smallest_vals(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erfc(val)};
-        const auto dec_res {static_cast<float_type>(erfc(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> second_smallest(1e-10, 0.5);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {second_smallest(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erfc(val)};
-        const auto dec_res {static_cast<float_type>(erfc(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> erfc_smallest(0.5, 1.5);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {erfc_smallest(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erfc(val)};
-        const auto dec_res {static_cast<float_type>(erfc(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> erfc_middle(1.5, 2.5);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {erfc_middle(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erfc(val)};
-        const auto dec_res {static_cast<float_type>(erfc(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 50))
-        {
-            std::cerr << "Float: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> erfc_large(2.5, 4.5);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {erfc_large(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erfc(val)};
-        const auto dec_res {static_cast<float_type>(erfc(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 100))
-        {
-            std::cerr << "  Val: " << val
-                      << "\nFloat: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> catch_all(4.5, 110.0);
-    for (std::size_t i {}; i < N / 6; ++i)
-    {
-        const auto val {catch_all(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erfc(val)};
-        const auto dec_res {static_cast<float_type>(erfc(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!(std::abs(float_res - dec_res) < 10*std::numeric_limits<float_type>::epsilon()))
-        {
-            std::cerr << "  Val: " << val
-                      << "\nFloat: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    // Negative values
-    std::uniform_real_distribution<float_type> negative_first_path(-2, -0.5);
-    for (std::size_t i {}; i < N / 8; ++i)
-    {
-        const auto val {negative_first_path(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "  Val: " << val
-                      << "\nFloat: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    std::uniform_real_distribution<float_type> negative_second_path(-0.5, 0.0);
-    for (std::size_t i {}; i < N / 8; ++i)
-    {
-        const auto val {negative_first_path(rng)};
-        const T dec_val {val};
-
-        const auto float_res {std::erf(val)};
-        const auto dec_res {static_cast<float_type>(erf(dec_val))};
-        const auto distance {boost::math::float_distance(float_res, dec_res)};
-
-        if (!BOOST_TEST(distance < 30))
-        {
-            std::cerr << "  Val: " << val
-                      << "\nFloat: " << float_res
-                      << "\n  Dec: " << dec_res
-                      << "\n Dist: " << distance << std::endl;
-        }
-    }
-
-    // Underflow case
-    BOOST_TEST_EQ(erfc(T{120}), T{0} * dist(rng));
-}
-
 int main()
 {
     test_fmax<decimal32>();
@@ -1913,16 +1603,6 @@ int main()
     test_nan<decimal64>();
     test_nan<decimal128>();
     #endif
-
-    // TOOO(mborland): specialize decimal128 to test extra ranges
-    // TODO(mborland): Move erf testing to its own file
-    test_erf<decimal32>();
-    test_erf<decimal64>();
-    test_erf<decimal128>();
-
-    test_erfc<decimal32>();
-    test_erfc<decimal64>();
-    test_erfc<decimal128>();
 
     return boost::report_errors();
 }
